requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
browser-history==0.4.1
build==1.2.2.post1
CacheControl==0.14.1
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
dotenv==0.9.9
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.16.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.12.1
grpcio==1.68.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keyring==24.3.1
Markdown==3.7
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
narwhals==1.37.1
newsapi-python==0.2.7
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
protobuf==5.29.0
ptyprocess==0.7.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pywin32-ctypes==0.2.3
pyzmq==26.2.0
RapidFuzz==3.11.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.9.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.45.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
toml==0.10.2
tomlkit==0.13.2
torch-tb-profiler==0.4.3
tornado==6.4.2
trove-classifiers==2024.10.21.16
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.28.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
